maint/build ~ (package) revise 'help' output to standardize visual whitespace
- use trailing "invisible" text to align help text on multiple platforms

## [notes]

The trailing text for vertical alignment is made using line(s) with a
simple '.' for low visibility (for cases in which the terminal does not
support ANSI escape codes). And then that text is made "invisible" by
using ANSI escapes which turn the character foreground color to 'black'
('\x1b[30m') and the text to 'hidden/invisible' ('\x1b[8m').

Unfortunately, `cmd` still adds an extra line prior to the prompt after
the help text, breaking platform visual consistency. This can be
mitigated by adding a character positioning ANSI escape sequence to the
end of the help text (`isWinOS && console.log('\x1b[2A')`). However,
this causes the last line of the help text to be overwritten by
the next prompt for WinOS PowerShell (and possibly other alternate
shells).

So, given that there's no standard way to detect the in-use WinOS shell
(ie, `cmd` vs `powershell`) and that `clink` [1] with 'flex-prompt' [2]
can be used to remove these extra lines produced by `cmd`, this
mitigation remains unused.

[1] [`clink`](https://chrisant996.github.io/clink)
[2] [`clink` flex-prompt](https://github.com/chrisant996/clink-flex-prompt)